app/requirements.txt
psycopg2-binary
matplotlib
pandas
numpy
nltk
wordcloud